lambdas/parseTransitData/requirements.txt
boto3
s3fs
pandas
psycopg[binary]
python-dotenv